aider: Create a new `dummydata.sql` file with SQL INSERT statements for the "Principles" table.
# Aider chat conversation:

USER: Create a database/dummydata.sql that we can run. For now, populate dummy principle items of "We are transparent" and "We are knowledgeable"

ASSISTANT: To create a `dummydata.sql` file with dummy principle items, we need to create a new file `database/dummydata.sql`:

1. Create a new `dummydata.sql` file with SQL INSERT statements for the "Principles" table.

Here is the *SEARCH/REPLACE* block:

database/dummydata.sql
<source>sql
<<<<<<< SEARCH
=======
INSERT INTO principles (principle) VALUES ("We are transparent");
INSERT INTO principles (principle) VALUES ("We are knowledgeable");
>>>>>>> REPLACE
</source>

This will create a new file `dummydata.sql` with SQL INSERT statements to add "We are transparent" and "We are knowledgeable" to the "Principles" table.