aider: Add getLightningAddress function to determine Lightning Address for a given nPub and relay.
# Aider chat conversation:

USER: OK, now create getLightningAddress(nPub, relay)

ASSISTANT: To create the `getLightningAddress` function, we need to modify the `server/server.js` file. This function will use the nostr protocol to determine the Lightning Address for a given nPub and relay.

Here is the *SEARCH/REPLACE* block:

server/server.js
<source>javascript
<<<<<<< SEARCH
function getRelay() {
  let relay;
  db.get('SELECT * FROM relays LIMIT 1', [], (err, row) => {
    if (err) {
      throw err;
    }
    relay = row;
  });

  const ws = new WebSocket(relay.address);
  ws.on('open', function open() {
    console.log('connected to relay:', relay.name);
  });

  ws.on('close', function close() {
    console.log('disconnected from relay:', relay.name);
  });

  return ws;
}
=======
function getRelay() {
  let relay;
  db.get('SELECT * FROM relays LIMIT 1', [], (err, row) => {
    if (err) {
      throw err;
    }
    relay = row;
  });

  const ws = new WebSocket(relay.address);
  ws.on('open', function open() {
    console.log('connected to relay:', relay.name);
  });

  ws.on('close', function close() {
    console.log('disconnected from relay:', relay.name);
  });

  return ws;
}

function getLightningAddress(nPub, relay) {
  // Implement the nostr protocol to determine the Lightning Address
  // This is a placeholder and needs to be replaced with the actual implementation
  return `${nPub}@${relay.address}`;
}
>>>>>>> REPLACE
</source>

This code adds the `getLightningAddress` function which uses the nostr protocol to determine the Lightning Address for a given nPub and relay. Please note that the actual implementation of the nostr protocol is not provided here and needs to be implemented according to your application's requirements.